LPS-97485 Split into interface and internal impl.
