+++ portal-kernel/src/com/liferay/portal/kernel/service/permission/ModelPermissionsFactory.java
@@ -16,6 +16,7 @@ package com.liferay.portal.kernel.service.permission;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.internal.service.permission.ModelPermissionsImpl;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Role;
@@ -68,7 +69,7 @@ public class ModelPermissionsFactory {
 		Map<String, String[]> modelPermissionsParameterMap, String className) {
 
 		if (className == null) {
-			className = ModelPermissions.RESOURCE_NAME_FIRST_RESOURCE;
+			className = ModelPermissionsImpl.RESOURCE_NAME_FIRST_RESOURCE;
 		}
 
 		ModelPermissions modelPermissions = null;
@@ -81,7 +82,7 @@ public class ModelPermissionsFactory {
 					CompanyThreadLocal.getCompanyId(), entry.getKey());
 
 				if (modelPermissions == null) {
-					modelPermissions = new ModelPermissions(className);
+					modelPermissions = new ModelPermissionsImpl(className);
 				}
 
 				modelPermissions.addRolePermissions(
@@ -115,7 +116,7 @@ public class ModelPermissionsFactory {
 	}
 
 	public static ModelPermissions create(String className) {
-		return new ModelPermissions(className);
+		return new ModelPermissionsImpl(className);
 	}
 
 	public static ModelPermissions create(
@@ -129,10 +130,10 @@ public class ModelPermissionsFactory {
 		String className) {
 
 		if (className == null) {
-			className = ModelPermissions.RESOURCE_NAME_FIRST_RESOURCE;
+			className = ModelPermissionsImpl.RESOURCE_NAME_FIRST_RESOURCE;
 		}
 
-		ModelPermissions modelPermissions = new ModelPermissions(className);
+		ModelPermissions modelPermissions = new ModelPermissionsImpl(className);
 
 		modelPermissions.addRolePermissions(
 			RoleConstants.PLACEHOLDER_DEFAULT_GROUP_ROLE, groupPermissions);
@@ -143,8 +144,8 @@ public class ModelPermissionsFactory {
 	}
 
 	public static ModelPermissions createForAllResources() {
-		return new ModelPermissions(
-			ModelPermissions.RESOURCE_NAME_ALL_RESOURCES);
+		return new ModelPermissionsImpl(
+			ModelPermissionsImpl.RESOURCE_NAME_ALL_RESOURCES);
 	}
 
 	public static ModelPermissions createWithDefaultPermissions(
@@ -154,7 +155,7 @@ public class ModelPermissionsFactory {
 			throw new NullPointerException("className is null");
 		}
 
-		ModelPermissions modelPermissions = new ModelPermissions(className);
+		ModelPermissions modelPermissions = new ModelPermissionsImpl(className);
 
 		List<String> modelResourceGroupDefaultActions =
 			ResourceActionsUtil.getModelResourceGroupDefaultActions(className);
